+++ CHANGES.rst
@@ -7,9 +7,11 @@
 
 - The Python implementation of ``Persistent`` and ``PickleCache`` now
   behave more similarly to the C implementation. In particular, the
-  Python version can now run the complete ZODB unit test suite.
+  Python version can now run the complete ZODB and ZEO test suites.
 
 - Fix the hashcode of the Python ``TimeStamp`` on 32-bit platforms.
+
+- Add support for Jython 2.7.
 
 
 4.0.9 (2015-04-08)